requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
numpy>=1.24.0
